Change MayI args behaviour to override environemnt instead of add args-actor
Robin is correct sometimes. Long ago, we discussed the behaviour of
envirnemnts, and how to ask the mayi library about things like number
of hosts and stuff. I wanted to keep MayI clean, but wanted be able and
allow actors to proved information.

And if we wanted to have a questions with values, like "Am I allowed to
save 53 hosts", the actors doesn't know about how many we want to save,
and thus we pass it as arguments, and it will end up as "args" varible
in the constraints

That sounds fair, but that means the constraints itself needs to handle
args, which Robin said. I don't rememeber exactly what he said
though... But of corurse there are better ways.

So lets change it. We can still have actors providiing the environemnt.
But if we change the arguments to be more like an "What If"-request.
"What if I have 53 hosts, am I then allowed to save?". The arguments is
then changed to an environemnt override variable, which can replace the
environment some variables, which makes the constraints handle the
request correcly without special cases, which reduces the risk of
regression errors in the future.

Change-Id: Id2b3debb4cb3975650ffa1cf07a483b4766b8763
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>