mayi: Debug-log when an action is denied
This helps troubleshooting when nothing seems to work. The log messages'
length is *huge* but useful, hence the debug level. We separate the much
shorter version that does not include env info to a notice level.

From what I can tell, we do not edit the log.yml file on upgrade to
include new fields, so I guess there's a manual process involved,
comparing .rpmnew files.

Change-Id: Ib2c824d46a20deda1886886598e257566adb0588
Signed-off-by: Carl Helmertz <chelmertz@op5.com>