helpers/sql: Replace generic "combine" helper with an OR and an AND helper
When there are 0 non-false cases, you'd expect an OR to exclude
everything (an OR is a list of valid cases - no valid cases, no
results), while an AND should include everything (an AND introduces
further restrictions - no restrictions means everything goes), but we
used AND's behavior for ORs as well.

Now there are two functions, and they both have the correct behavior
for the empty case. With this in place, bug #8149 should be resolved,
for real this time.

Change-Id: Ibdaac7a24147c0fe304da3b930cc8f77435aed12
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>