avail: Implement option for rendering time
We used to always unconditionally render the time and the percentages
for single-object reports, and only ever render percentages for
multi-object reports.

For me as a developer, being able to extract the time as well as the
percentage is very, very useful - in bug #8932, I wouldn't have had to
guess about the cause, I could have just turned it on and quickly
verified it.

And if I was a user, I want percentages for talking about SLAs, but I
want absolute time to talk about impact, and I can't say off the top
of my head how long 1% downtime during workhours in the last month
actually is.

Change-Id: Idef80e28585e21b37afdae32082ccd3001d52daa
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>