MayI controlled top-left logo branding
Message of MayI query is used as for branding,
if the message maps to an image in the brands
folder that image will be used for branding.

Otherwise a generic op5 banana is displayed
followed by the message.

This introduces the possibility of new "license-types"
before there is an actual new logo for it, that is, if
the message contains "Bananarama PoC".

It will render the brand close to what the official
logos look like until such a time as a new release contains
a proper logo for it.

Edit: Updated to a branding helper.

Edit: Updated, moved license brandings to license module

The current icons that exist are:

 - enterprise
 - pro
 - trial
 - free
 - op5 (generic default)

Change-Id: I9b0301beb73fae8628493ef2049f8e25f3ad3224
Signed-off-by: Tobias Sjondin <tsjondin@op5.com>