Make modules override application, instead of application override modules
Once upon a time, modules where made to be a resource for the
application to access and use. Application could then include modules
but override classes and variables. Nowadays, ninja, the application
with Kohana behind, is a framework to handle modules. Thus modules is
made to change the behaviour of ninja, making it possible to override
the ninja behaviour.

Thus the order of inclusion should be reversed, lowest priority is the
system, then the application, and above that the application. Therefore
we need to change the order of the include path

This makes it possible to override the default logo with a branded one
when installing a module, by just providing an image with the same
name, if the image is located with Kohana::find_file

Change-Id: I029cbf539a7af5d1c99ddf182a2d77dea8a74174
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>