Login redirects to requested URI
If you request a URI but you are not logged in;
add the requested uri to a get parameter and redirect
to this once login completes.

If you have another window open at login page and
reload it, it will if a uri parameter is given redirect
to that uri, otherwise it will redirect to the default
page.

Tested with some generic URIs and a long-ass
report generation URI and it seemingly works fine.

Updated to use Router instead of URI library, seemingly the
end result is the same in the above tests I made.

Fixes: #9245

Change-Id: Ie16c1eb7bb31274063ef381986030517a9218de0
Signed-off-by: Tobias Sjondin <tsjondin@op5.com>