widget_model: Don't copy default widgets when saving globally
The fetch all widgets method returns all widgets that the user owns AND
whenever there is a row where the username is NULL. The username NULL
means it is a default template for a widget.

Problem is it's instance_id is also NULL since it is not used anywhere
making the save function think it should create a new instance of the
template with the new settings that were saved merged into it.

By ignoring any widget object with an instance_id of NULL we can make
sure that refresh_interval can be saved on every widget for the user
whitout causing a regression with hidden widgets reappearing after the
save.

This fixes #9380 and keeps functionality for #7082

Change-Id: Ie5185e93a2dfd8a96b098d41417b746d36d473dc
Signed-off-by: Henrik Smederod <henrik@smederod.com>