Add nagbar messages to print notifications
Since licenses should be able to output messages to the reports too, we
should print the messages to print_notifiations also.

MayI and the new license interface restricts messgaes depending on
actions instead of where to put the messages, and thus, if it should go
to print or to display, should be controlled by the license depending
on action instead.

This resolves #9149, given that report is controlled by MayI (#8867)

Change-Id: I9081c0a3e7432eb6fb0a156dd6e91befd7553e21
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>