Make reports use MayI access control, assume hosts though
Since reports has lot of entry points, it's really hard to track
authorization correctly. But we need to be able to control access to
reports in general.

This patch makes it possible to control permission to the entire report
subsystem, but not specific parts. It handles autorization in the
controller constructor, which isn't that good either, but works for now.

To resolve this correctly, it would be a week or so, and should
probably be done when rewriting reports to use ORM sets.

Since the controller constructor yet doesn't know what kind of objects
reports should be created on, assume it's hosts, which should be quite
ok for now.

Resolves #8867

Change-Id: Id1ea2e4483e27fd212f04ca73905613185b00aae
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>