httpapievent: Make state selection a lot more useful
First, convert the field into an array. Like most other arrays in our
API, we still support a non-array notation, but we also support
multiple options. This means you can select
host_states[]=up&host_states[]=down to get all up and down, but not unreachable.

Second, add the option "none". This means you can select no host
states at all, no service states at all, but include downtime information.

Third, for the whole state name conversion to work properly, we need
to convert the name to a value before the rename stuff kicks in -
sure, we could decide that the API is a special case and that we
sometimes receive keys instead of values and we should just be content
with either, but it's nasty and I really don't like it - the rename
stuff in Report_options certainly expect option values, not names, so
the HTTP API should provide that to be consistent and avoid bugs.

This implements the adjustments as requested by #7968.

Change-Id: I62ea323ffd3f8f563c591231cd0d008713bebd0f
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>