view_config: Tighten xss
I am not sure why I could not use addcslashes(str, '"\'') since the
resulting html looked OK but the browser still chose to render the XSS'd
HTML. I blame the multitude of layers of passing around HTML as strings
instead of constructing the DOM with nodes and methods to control
attributes and values. Hacklang, anyone?

This fixes #9415.

Change-Id: I30f04ffca8198b9ed81b4f0545c426de3731448f
Signed-off-by: Carl Helmertz <chelmertz@op5.com>