Branding in reports
Added branding functionality to report footer.
Wish we didn't "have to do" it this way, but anyways.

It displays rather poorly and I'll look into that, don't
really know why at the moment (image quality that is).

A guess would be printing css.

In the case of labeled branding, i.e. op5 banana and a label.
it will omit the label due to improper css and all current
monitor systems to comtinue displaying reports the same way.

Also fixed may_i namespacing

Completes: #8854

Change-Id: I137abd9ef4c19da9081d4aa4e648c2ef78dd7523
Signed-off-by: Tobias Sjondin <tsjondin@op5.com>