More debug output when autoloading fails outside production
If autoloading triggers, a class is needed to be resolved in order for
the system to work, otherwise a class not found error will trigger. So
in fact, if autoloading fails, sometihing is wrong.

If not in production, it would be nice to have debug output, like
include paths, types and resolved filenames. So if not in production,
throw an exception containing that information.

Change-Id: I2dad73760061e4ef4dfcd3d7642351ac528bc564
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>