Convert user_mayi_authorization to implement ACL:s
Since API needs to be added to user_mayi_authorization as negation if
not allowed, we need to have a more advanced ACL implmenentation for
user_mayi_authorization. That means all the old lists of what should be
allowed needs to be converted to ACLs, and api_ needs to be added
before as disable access to API-namespaces if api is requestesd.

That means, you need both api access and hosts access to access hosts
through api

Change-Id: I5fe2da79252dbb5d3a67a3c5586210efb62171b4
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>