Problem count; Stop blinking!
The quicklinks are made up of a link with shields as background and a
span that displays number of problem objects. Whenever an update to the
link occurs the span is removed and created again with the new number of
problems.

Instead of removing it update it's text as to avoid blinking if the
request is slow.

Additional fix to #8312

Change-Id: I8a6a068086169f725cb0e46eb68694ac6d4ca9a2
Signed-off-by: Henrik Smederod <henrik@smederod.com>