Load information for sysinfo in constructor
Was a little bit to fast when reviewing the previous change about
op5sysinfo lazy loading. Thus I'll have to fix it now.

Since op5sysinfo is a singleton, and loading should be done when
instanced, it should be done in the constructor, so move it to the
constructor, not the instance method, which fetches the current
instance.

Change-Id: Ia8aaed8812d5e36c66756f52815f7408c83f45fd
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>