Temporary move nacoma hooks to different name
Since we want to move nacoma to a module in ninja, we want to make sure
we don't have naming conflicts.

Ninja contains an old hook digging through some nacoma information.
temporarly move that to a seperate module, so it can be merge to the
nacoma repository at a later point.

Change-Id: I644a1d862755574346f72a4ee9dc87e942b71077
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>