mayi: Always allow accessing listing saved reports
Parts involved:

The pre-2015.albino way to list saved reports was to enter through the
"create a report" interface. That list is determined by host_view_all =>
list all; else => list only mine. See Report_options.php.

Now that we have exposed it as a list view, we got the "list only mine"
logic implemented through the ORM-sets' get_auth_filter() method (see
modules/reports/models/savedreportset.php). But, to even come that far,
we need to authorize access to the listview-table. Listviews asks the
ORM-model's get_mayi_resource() and found monitor.reports.saved_reports
as a resource, which we never ever gave access to. That's where this
patch comes in.

We grant access to all actions under that namespace to not cause any
regressions and to keep things as uncomplicated as possible. There's
nothing else than ":read.list" at the moment.

This fixes #9762

Change-Id: Id1bbd16a1ae39009e9f701d7fa59210bad0cd30d
Signed-off-by: Carl Helmertz <chelmertz@op5.com>