views/auth: Tell the user why they're not allowed
"You don't have permission to view this page" gives me absolutely
no hint of how I'm supposed to gain access to <this page>. This patch
makes sure that the user is informed of just that, and if no such reason
is provided urges them to file a bug report! All in big, friendly letters
and a polite, personal demeanor.

This fixes #9789

Change-Id: I636b6c3262e957a951825aed7539e38790773829
Signed-off-by: Anton Lofgren <alofgren@op5.com>