mayi_auth: Force session on populate_mayi
Due to initialisation ordering reasons, we sometimes would not have an
active sesion when populating mayi, which would lead to the user not
being authenticated properly. This patch forces the session, which takes
care of the problem short-term. MS suggests refactoring op5auth to
handle sessions ubiquitously in the future.

Change-Id: I9eca51baba20cc72ddd24f24ee3a9693234e8da6
Signed-off-by: Anton Lofgren <alofgren@op5.com>