mayi: Add more translations for authpoints used in Nacoma
We missed all kinds of objects, such as escalations, manpacks and some
templates. There are some more hanging around, such as "export", which
does not have a suitable 1:1 mapping to anything that has to do with
objects, at least not if you take into account that "create" means
"POST" to Nacoma, and "export" means "take Nacoma's changes in its db
and put them in config files".

Change-Id: I09d351074aad113eb795b233a154429d9a31af14
Signed-off-by: Carl Helmertz <chelmertz@op5.com>