session: Keep Session::instance() interface for CLI
In 365f869365b20523ed16d9a78194eb703c4ec7ec, we just wanted to avoid the
session_start() call so all logic after "are we CLI? abort!" was
ignored. That was very sloppy, since other code calling the CLI might
contain calls like "Session::instance()->get('x')". That did not work
too well, since we never set self::$instance, enabling the glorious and
almighty singleton pattern.

So, this patch basically mocks out all function calls that are part of
the API, and pretend (in maaany places) that the session is empty (which
it is).

This fixes #9767 once again.

Change-Id: I115e8032de820a6f9039f283df6baf1b90f816c4
Signed-off-by: Carl Helmertz <chelmertz@op5.com>