stacktrace: Suffix the dump with html
This fixes MON-6851 and MON-7509 by adding the html suffix and naming
the file with a 24-hour hour time part as well. The meta tag declaring
UTF-8 works well with elinks, lynx however does not react to it. Firefox
shows it nicely, of course.

Using sha1() instead of the previous tempnam() solution creates a much
longer filename, fyi.

Easiest way to verify:
- set IN_PRODUCTION to 1 in ninja/index.php
- open ninja/application/controllers/ninja.php and add the code 'throw
  new Exception("räksmörgås");' within the constructor (after
  parent::construct())
- load any page in ninja
- yum install elinks
- elinks <the path that was shown in the browser>

Change-Id: I76345a008618cf6975b6b437cf96a8dbad1a2af8
Signed-off-by: Carl Helmertz <chelmertz@op5.com>