PNP graphs: Don't include_once if it isn't class-files to load
When loading config files, which is a php file modifying an array, it
shouldn't be included once.

It works in this case when loading from web, since it caches the first
execution in session. But thorugh cli, sessions are disabled, and thus
no caching is done, and thus only the first load will result in a valid
$conf array

Resolves: MON-7942

Change-Id: I1787804456c3cf97620bec86380fbdb03070c78a
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>