Fix PNP links in listviews
Hovering the PNP graph links in listview should create a
popover which displays the source=0 and view=0 PNP graph.
However, in internet explorer the target of the event of
an element load is not the loaded element (who knew) but
the document element (possibly since that was the one
who loaded the element?), anyways, changing the src
attribute of a htmldocument does what you would expect,
change the source of the entire page, i.e. the location.
Which whites out the screen and fails everything.

Fixes: MON-7967

Change-Id: I75a7690d62a195d35ede1f160b11c48985461f0e
Signed-off-by: Tobias Sjondin <tsjondin@op5.com>