External widgets: use correct op5auth instance
Well... the op5auth::factory does validate the user, and creates a new
op5auth instance correctly authenticated. But...

If the op5auth instance is replaced, the op5auth instance registered as
a mayi actor will be invalid. The correct way is to update the current
op5auth instance to the correct user, but disable session writeback.

During fixing, also a minor favicon-issue were found in the external
widgets, that is fixed

Resolves: MON-7974

Change-Id: I3f3020e56535e512f4ac18512b7e1d0704201af4
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>