recurring downtime fixed duration and commands
When the new object selector came into place this
view was ignored and did not recieve proper parameters
for the nagios command string and therefor failed the
command, this has been fixed.

Duration has never been handled properly, the field states
hh:mm:ss but simply inserted it as a string into the command,
therefor it only worked it you input a int/float value up until
some time after 6.2.1, after that we include a check for the
format, but still didn't allow the format. Good job.

Fixes: MON-7301 MON-6519

Change-Id: I72d9ff327d64f4d8b77f7f1fc53b4a25c3a64fb6
Signed-off-by: Tobias Sjondin <tsjondin@op5.com>