Initial commit of js-bundling
Using "manifest.js" files to dictate what should be contained
in the bundle, we should move away from this and adapt to the
commonJS pattern within the source files, but as long as we
are heavily dependant on dated jquery plugins and an old jquery
this is close to impossible.

Anyways, this is a test of browserify and performance improvement
of bundled JS files.

Change-Id: I1b8a48e5ef36b48d7e7d7f7a5fb7986a469eeb52
Signed-off-by: Tobias Sjondin <tsjondin@op5.com>