db_migrations: Wrong path to previously installed files
I just merged maint/7.0 into maint/2015.albino. Git was smart enough to
know that a file was moved, and applied the diff on the newly moved
file. The problem is that the diff was a string representing a path
which no longer was correct, since it was relative to the file itself.
Let's adapt that path to the file's new location.

Change-Id: I05725fde267e51f30aa6ec749bc3fc55ae643820
Signed-off-by: Carl Helmertz <chelmertz@op5.com>