auth: Close session on write_close
Keeping the session open for too long blocks other, concurrent,
requests. Releasing the session sooner allows for better parallellism,
which in turn leads to a better (perceived) performance from an end-user
perspective. For example, this allows multiple list views to be
rendered asynchronously, as opposed to having to wait for all of the
requests to finish before rendering.

This fixes MON-8031.

Change-Id: Ida8fcf05f989fb08edb659b31fadf23a36c58224
Signed-off-by: Anton Lofgren <alofgren@op5.com>