auth: Force new CSRF token on login
Validating the CSRF token when we're about to login makes little sense,
and only causes problems. When the user is authenticated, there is no
reason not to simply generate a new token for the newly created session,
regardless of whatever leftovers might be present in the session.

This fixes MON-7598.

Change-Id: Iecc0cce7b0fd37992bca5afc54d6c02c18829cee
Signed-off-by: Anton Lofgren <alofgren@op5.com>