random: Don't intentionally reduce entropy
This whole function (except for actually using mt_rand) is a mess.

Get rid of the downright harmful parts, parts we don't need or use, and
stop trying to be clever.

The set of strings 'a' through 'z' is a subset of the set of the
alphanumeric strings, and so is the set of digits '0'-'9', which means
we've already *got* an alphanumeric string. Reducing randomness in order
to "look" random is a horrible idea.

Change-Id: I9f58783d7a625241196192f6599a8de9dca590ae
Signed-off-by: Anton Lofgren <alofgren@op5.com>