httpd.conf: Enable KeepAlive
Several of the pages in Ninja make a lot of requests to fetch various
resources, such as CSS or JavaScript. Having to set up and teardown the
HTTP connection for each resource adds a significant overhead to the
loading of Ninja, which results in longer waiting time, and a worse user
experience.

Enabling KeepAlive should thus provide a significant, palpable
improvement to user experience.

This fixes MON-8042.

Change-Id: I1c0b9c097a683097b23b7fb33eba944aba434012
Signed-off-by: Anton Lofgren <alofgren@op5.com>