search: Improve error message when no result
Since the original query might for example be:

"h:monitor AND sg:monitor AND l:monitor"

It's kind of ugly and unexpected to print that in the search result
listview for service groups, especially when both hosts and log messages
have matches. Besides, I already know what I searched for, I want to see
how it was interpreted for the particular thing that wasn't found.

Change-Id: Ib8f56375a48660a59c834ba12c6e3804cb11fb36
Signed-off-by: Anton Lofgren <alofgren@op5.com>