Revert "auth/janitor: Destroy session on logout"
This reverts commit a0f4a98eb3546892d72d3f17c9c3974eabe24a99.

In short: with the previous patch, nobody could login to ninja.

We can't destroy the session on logout() since login() (and other
methods if they wish) calls logout() before doing anything else, to just
get a clean state of the session. Importantly, the session is to be kept
open. If we destroy the session and then try to set the "yeah, you are
logged in now" information, we can't, since the session is not reopened.
I don't think it should be reopened either, it messes with the lock that
is made up by session_start() and session_write_close().

Change-Id: I649f931e226b51218188de6703c08688112c332f