popover: Fixed abort-state "Failed" issue
When a popover was briefly hovered and then left
it would abort the AJAX request, which in jQuery
will execute the error handler which in turn displays
the "Failed" message. It now checks if the status is
"abort" in which case it does not display a message.

This applies to help:, get: and post: prefixes of
the popover attribute.

I also added some better messages on actual failures,
such as a graph, helptext etc. could actually not be found,
atleast better than "Failed."

Fixes: MON-8091

Change-Id: Id32404d734c657a3861f942be90184d83b7e4c4c
Signed-off-by: Tobias Sjondin <tsjondin@op5.com>