Let views render themselfs
Kohana supported that views can access controller variables through
$this->xx, beside the parameters we pass through the template.

There were several unnessecary references to $this, which also is
resolved.

Allowing the view to access controller variables means the view depened
on the current controller, and views becomes really tight dependent on
the controller, which is strange.

It also means that views can't be rendered before the controller is
created, and thus no error page is possible to render in the event
system.pre_controller or earlier.

Just let the view render itself, by move the code to render the view to
the view itself.

Fixes: MON-8129

Change-Id: I58bdd9282ef1cce9d2a66e6143414751dc875f2e
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>