csrf: Remove expiry stupidity
The CSRF token should not have its own timestamp or expiry time. It
should rely completely and utterly on the session to which it is tied.
And the session already has its own expiry logic associated with it.

To quote OWASP's general recommendation for synchronizer tokens:

"In general, developers need only generate this token once for the
current session. After initial generation of this token, the value is
stored in the session and is utilized for each subsequent request until
the session expires."

Change-Id: I671ab01bd85f158b5608842fa37e90715630f5fb
Signed-off-by: Anton Lofgren <alofgren@op5.com>