csrf: Validate CSRF token on all POST'y requests
We communicate denial by 403'ing, which should send a pretty clear
message that the user's doing something she's not supposed to.

Now, the 403/404 pages look pretty bad, so we might want to do something
about those as well, but that's out of scope for this particular patch.

This fixes MON-8082.

Change-Id: I4f5a46e20989a2118420295014837b49d9b773fe
Signed-off-by: Anton Lofgren <alofgren@op5.com>