auth: Destroy session on logout
This was reverted once already, with the motivation that:
"We can't destroy the session on logout() since login() (and other
methods if they wish) calls logout() before doing anything else, to just
get a clean state of the session."

Of course, the solution isn't to not destroy the session, the solution
is to stop frickin' logging out when what you should be doing is logging
in.

I believe I've convinced Carl of the righteousness of this too, this
time around. And I know TS agrees.

One caveat; we need to do "$this->session_clear()" as well, here,
because the unit tests currently rely on that happening.

Change-Id: I1d64f488f1afbebd0030cf2347222a6109173a60
Signed-off-by: Anton Lofgren <alofgren@op5.com>