Explicitly tie CSRF token to session
Kohana's session handling is quite insane, but its CSRF helpers are
worse. So, let the lesser of two evils prevail, and get rid of the CSRF
class completely and move what little logic was left in there into the
session handling layers; that's where the logic belongs regardless
(which can be seen by looking at the coupling to session state in the
csrf class).

This gives us a clear, explicit, no-frills 1:1 mapping from a CSRF token
mapping to an active session, which should make it much harder to
accidentally reset the CSRF token for an active session, something that
happened (figuratively) all the time previously.

In addition, this commit moves the CSRF config to system/ since this
enables PNP to find it as well as Ninja - which it needs to because
the CSRF logic now resides inside Kohana core.

Change-Id: I0fd0746e8c38d9bc97cf5829181e9f1cc147e171
Signed-off-by: Anton Lofgren <alofgren@op5.com>