orm: Add naemon commands to objects
All Naemon commands are defined 1:1 in structure.php, mapping to methods
in respective ORM objects, so that Host_Model::schedule_check() now
exists for example. The names are arbitrary and (as long as they're not
used in the rest of Ninja yet) can be changed there.

Errors are reported both as a boolean return code and the output from
the command as a reference variable, as our customs dictates (see
op5mayi::run() for example).

These auto generated methods gives us the old, ugly way of running
commands, but in a structured way with methods and arguments instead of
"important strings" everywhere. What's even better is that we now have a
good common ground to build our "everyday commands" on. I've included
one such example, Host_Model::check_now().

All authorization is handled outside of the commands. Since each command
exist on an object instance[1], you won't have that object unless you're
authorized for it. If you want to apply mayi logic to it, there's a new
command on each model called get_naemon_commands() that gives you all of
the object's methods that triggers commands.

This implements MON-6761.

1: You no longer pass in host_name, you
HostPool_Model::all()->reduce_by('name', $name, '=')->check_now();

Change-Id: I4c1b198f4ed44860c2ca560d4d4d623480557d5b
Signed-off-by: Carl Helmertz <chelmertz@op5.com>