spec: Don't try to "install" widgets if they are already "installed"
We always try to "install" widgets during installation/upgrade, but
there is no need to do this if we already have done it in the past.

So now we check if we have a match of the widget before we try to
install it in our system.

This fixes MON-8183

Change-Id: Ida24ca8d7128d66ceb288d2904b923bd22f012eb
Signed-off-by: Mattias Ryrlén <mattiasr@op5.com>