view: Make PNP's views able to refer to $this->variable
A Ninja commit[1] removed the possibility for views to share its
controller's properties, accessed like `echo $this->foo;`. PNP reuses
Ninja's system folder/library and thus got this change for free. Now it
turns out that Ninja had to replace `$this->foo` with `$foo` in
templates, but that jobs was not done in PNP. Since we want to be able
to merge with upstream PNP in the future, we should keep the `$this->`
convention for that specific reason.

This only applies for PNP's controllers, indicated by them inheriting
PNP's "base controller" System_Controller. Ninja has one controller like
that on its own, but it has not been touched since 2009 and 5/5 people
that happened to be at the office right now, had not heard of it. So,
let's delete it in order to match on that class name.

This is a part of the fix for MON-8185, the other part is a commit in
PNP.

    1: 638b59ab6b132c305cb9c8c2e50c1c31bd99bffe

Change-Id: I85af80f3edbc7f4d27a8238aefc0dc4f55950a40
Signed-off-by: Carl Helmertz <chelmertz@op5.com>