report: Expose remove_object() and rename_object() in Report_options
These methods were implemented in the CLI controller but are now inside
the report options model itself since the logic really should not be
exposed outside of the model. With that said, this patch only exposes
some helpful methods to deal with object being renamed/removed, it does
not expose less information than before.

Besides moving code around and adding tests, we also solve some issues
with inheriting ArrayAccess: $access_instance[0][3] = 4; does not work,
since $access_instance[0] is translated to
$access_instance->offsetGet(0) which does not return a reference to that
array, but a copy. Thus, any changes to it are ignored.

This fixes MON-7190.

Change-Id: Idced7a68e33f3ece276218079242517665fa6add
Signed-off-by: Carl Helmertz <chelmertz@op5.com>