recurring_downtime: Submit form to recurring_downtime controller
Errors:
* form::open() to wrong url
* form::hidden() never echoed
* no title tags on anchors in listview
* no csrf token

form::open() equals <form action="">
form::open('') equals <form action="/index.php/monitor">

Kohana is to blame, in an upcoming commit I will change the behavior of
the second row to become the first one.

Besides, the return value of form::hidden() was never printed so editing
a recurring downtime would not work even if we posted data to the
correct controller.

Since the cucumber step finds titles this way:
a[title=X]
and not:
a span[title=X]
we also need to add the title attribute directly on icons' surrounding
anchor elements.

This fixes MON-8223

Change-Id: I600047c2f8cda9a4a9007dc4b58253eae04e97ca
Signed-off-by: Carl Helmertz <chelmertz@op5.com>