change_password: Don't always print out empty alert box
isset($always_set_variable) is not a smart thing to check for. On the
flip side, the controller logic is nicely written regarding the fact
that it initializes the variables it uses!

Also, a toast to Tobias, with whom I will fight many style-battles in
the future. Cheers!

This fixes MON-8224

Change-Id: I9662a2bbfcfe53dcf1bae57801a4d411e8bff2ff
Signed-off-by: Carl Helmertz <chelmertz@op5.com>