form-helper: Make sure form::open('') == form::open()
The syntactic confusion is especially annoying when there are more
parameters to the form::open() call, requiring the first parameter to be
"something".

This will lessen the possibility of bugs like MON-8223 in the future
(well, that, and that we test our code a tiny bit better).

The only place in Ninja that uses there parameter less form of
form::open(), is the change-password page. There never was a test of its
functionality so I added one.

Change-Id: Ie63749efac20989e42c6744d0214910b76924d03
Signed-off-by: Carl Helmertz <chelmertz@op5.com>