Move menu to module, introduce nested events
When an error occurs we should not continue, this if prevented by the exit(13)
(13?????) but if exited the post_controller would never run and the menu object
would not exist to be rendered.

Though, the error controller inherited from ninja controller and therefor
emitted the ninja.menu.setup event which overwrote the event data of the
erroring event and an error was never seen and you got the route logic instead.

So we moved the menu to a module that hooks into the post controller event so
that the error event is not overwritten when the ninja controller constructor
is run. The ninja controller instead draws the menu reference from the menu
setup event.

The ninja.menu.setup event does not happen on error pages (since the post
controller is not run), so the Error controller instantiates a new menu object
which is empty.

This in combination with previous commits results in CSRF working but the error
page does not have a proper menu.

This fixes MON-8082

Change-Id: Ia47b2219312ce878592ca39231c4e51023866c9e
Signed-off-by: Tobias Sjondin <tsjondin@op5.com>