navigation: Use correct path for nagvis
Nagvis does not exist under "nagvis/nagvis" but under "nagvis", corrected this,
why was this not found before? It is probable that this happens due to my fix
that prohibits further execution on http error events (fixing 403 forbidden)
since the nagvis controller has no nagvis method but a catch-all __call which
is not found by the router when determining if the route exists but is
executable if it is allowed to proceed (resulting in a 404 -> but hey lets
execute this method).

This fixes MON-8247

Change-Id: I5459eb8d7d0912e0c0f76ad97f6054ca02c130d5
Signed-off-by: Tobias Sjondin <tsjondin@op5.com>