Start using ORM for naemon commands
Part of MON-8262

This creates a structure for the general view/controller logic that
replaces the old command-controllers since the old think was 1:1 to
Naemon commands, now we're N:1 between instances-of-ORM-object-class and
command (where a "command" is a method on said instance).

The controller makes it easy to implement MON-8263 and MON-5930, by
essentially copying Exec_cmd_Controller::obj() to something like
Exec_cmd_Controller::filter() that takes a filter as input, instead of a
table + ḱey pair.

Change-Id: I730081730b71ff0ced8fdfd7adddc4b8bc1e6b22
Signed-off-by: Carl Helmertz <chelmertz@op5.com>