backup: Remove duplicate .tar.gz ending from filename, add link titles
The method Backup_Controller::download() takes a filename, not a
filename-without-extension. It even finds the complete filename on disk,
before adding the superfluous file extension.

The icons' meanings are very hard to parse, let's at least add title
attributes on them so that the user can hover the links to get a hint.

This fixes MON-7568.

Change-Id: Iba5297beceb0f29c61ddfc5c40d66bb44cacd5ae
Signed-off-by: Carl Helmertz <chelmertz@op5.com>