Global search result page no longer reloads after 90 seconds.
According to MON-8237 the auto-refresh of the search result page may cause Logger search result to be
empty for a while when new data is fetched. I couldn't reproduce this behavior but I disabled the page
auto-refresh, and checked that the page doesn't reload by itself anymore.

This fixes MON-8237

Change-Id: I8f58d6eed5c519d5e7e2f8cf8c081a9de860eed7