Refactor: Builders, keep source in module directory
And also generalize the generators. This simpleifies generating and
triggering builers really much, which makes it easier to implement and
integrate new builders.

Simple builders which would help a lot is docstring => manifest
builders, which makes it possible to tag functions in models, so
commands and such can be exported. Also methods can be triggered with
security-tags, like mayi resources and such.

Resolves ninja part of MON-8265

Change-Id: I84a2fcf3caaec2e0b0e777c854141973269e8431
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>