Fixes filter search in config.
Filter do now work on all objects types that can be selected
on the page. Previously it assumed that the object type had
a "name" field (which isn't true for e.g. Services).
Also changed the placeholder text "Enter text to filter" to
a "modern" html5 placeholder attribute and remove javascript
etc. that was needed for it to work (which it didn't) earlier.

This fixes MON-6993.

Change-Id: Id23e7d2528d08ada323c4b6c352739ffe734b764