command: Move group commands from controllers/views to ORM
We modify the interface for submitting scheduled downtimes in the past,
to take an optional ORM set as a first parameter.

Because of that, we don't have to pass around objects all over until it
is time to actually use the object names, for insertion into the report
data. An exception to this rule is the newly introduced
ServiceGroup_Model::get_hosts_set() which actually needs to fetch all
hosts up front, causing possibly bad performance.

Change-Id: I3ce4bb85c5ca781dd999e7020cbd493b7487e017
Signed-off-by: Carl Helmertz <chelmertz@op5.com>