Implement generic multi-action + minor fixes
Multi actions hasn't changed almost anything the last couple of years,
but many thoughts about it has come since then. But commands was just
to horrible...

Then the new command interface came to life, originating from objects
in ORM, and objects grouped themself in sets, and... a new mutli-action
concept was born.

However, libraries are in constant development, and some parts in ORM
was missing for the newborn multi-action to actually grow to a fully
fledged feature, like some generic methods to know who the object is
(class_obj, class_set, class_pool, get_table for all objects, sets and
pools). Also objects sometimes want so be referred to as groups, even
though they are alone (pool->set_by_key()).

But what is an awsome newborn feature, if it can't be accessed, so the
list view wan't happy anymore. It screamed for care, so the multi
action needed a new menu, to show of all the features, and to dim down
objects that doesn't want to take any action.

And voilá, we got a new cool feature!

Multi action part of MON-6761
Multi action part of MON-5259
Resolves MON-5384
Resolves MON-5511
Resolves MON-6472
Resolves MON-6477
Resolves MON-6902
Resolves MON-6720
Resolves MON-6760
Resolves MON-7082
Getting really close to MON-5930

Change-Id: Ie0188ef3f495d03ddae87dc1cd181cf83d41c8ba
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>